+++ paper_rq/models.py
@@ -1,4 +1,4 @@
-import inspect
+import datetime
 import logging
 
 from django.db import models
@@ -12,7 +12,7 @@ from rq.job import Job, JobStatus
 from rq.queue import Queue
 from rq.worker import Worker
 
-from .helpers import get_all_jobs, get_all_workers, get_job
+from . import helpers
 from .list_queryset import ListQuerySet
 
 
@@ -73,7 +73,7 @@ class QueueModel(models.Model):
 class WorkerManager(BaseManager):
     def all(self):
         workers = ListQuerySet(self.model)
-        for worker in get_all_workers():
+        for worker in helpers.get_all_workers():
             obj = self.model.from_worker(worker)
             workers.append(obj)
 
@@ -85,7 +85,7 @@ class WorkerManager(BaseManager):
             pk = kwargs.pop("name", None)
 
         if pk is not None:
-            for worker in get_all_workers():
+            for worker in helpers.get_all_workers():
                 if worker.name == pk:
                     return self.model.from_worker(worker)
 
@@ -117,7 +117,7 @@ class WorkerModel(models.Model):
 
     @cached_property
     def worker(self) -> Worker:
-        for worker in get_all_workers():
+        for worker in helpers.get_all_workers():
             if worker.name == self.name:
                 return worker
 
@@ -129,7 +129,7 @@ class WorkerModel(models.Model):
 class JobManager(BaseManager):
     def all(self):
         jobs = ListQuerySet(self.model)
-        for job in get_all_jobs():
+        for job in helpers.get_all_jobs():
             try:
                 obj = self.model.from_job(job)
             except DeserializationError:
@@ -146,7 +146,7 @@ class JobManager(BaseManager):
             pk = kwargs.pop("id", None)
 
         if pk is not None:
-            job = get_job(pk)
+            job = helpers.get_job(pk)
             if job is not None:
                 return self.model.from_job(job)
 
@@ -166,6 +166,9 @@ class JobModel(models.Model):
     enqueued_at = models.DateTimeField(_("enqueued at"), null=True)
     ended_at = models.DateTimeField(_("ended at"), null=True)
 
+    # флаг, устанавливаемый при ошибках десериализации задач
+    invalid = models.BooleanField(_("invalid"), default=False, editable=False)
+
     objects = JobManager()
 
     class Meta:
@@ -178,38 +181,39 @@ class JobModel(models.Model):
 
     @classmethod
     def from_job(cls, job):
-        if job.instance:
-            if inspect.isclass(job.instance):
-                instance_class = job.instance
-            else:
-                instance_class = job.instance.__class__
+        invalid = False
 
-            callable = "{}.{}.{}".format(
-                instance_class.__module__,
-                instance_class.__qualname__,
-                job.get_call_string()
-            )
+        try:
+            job._deserialize_data()
+        except DeserializationError:
+            invalid = True
+            job_callable = None
         else:
-            callable = job.get_call_string()
+            job_callable = helpers.get_job_func_repr(job)
 
         return cls(
             id=job.id,
             queue=job.origin,
             description=job.description,
             timeout=_("Infinite") if job.timeout is None else str(job.timeout),
-            callable=callable,
+            callable=job_callable,
             result=job.result,
             exception=job.exc_info,
             meta=job.meta,
             created_at=job.created_at,
             enqueued_at=job.enqueued_at,
             ended_at=job.ended_at,
+            invalid=invalid
         )
 
     @cached_property
     def job(self) -> Job:
-        return get_job(self.id)
+        return helpers.get_job(self.id)
 
     @property
     def status(self):
         return JobStatus(self.job.get_status(refresh=False))
+
+    @property
+    def enqueue_time(self):
+        return self.enqueued_at or datetime.datetime(datetime.MINYEAR, 1, 1)