+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.5.5](https://github.com/dldevinc/paper-rq/tree/v0.5.5) - 2022-04-19
+### Features
+- Display job function.
+- Display jobs that raises `DeserializationError`.
+- Changed default job ordering.
+### Bug Fixes
+- Fixed an issue with requeuing scheduled jobs. 
+
 ## [0.5.4](https://github.com/dldevinc/paper-rq/tree/v0.5.4) - 2022-04-12
 ### Features
 - Add custom scheduler class.